backend/requirements.txt
alt-profanity-check==1.7.2
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
defusedxml==0.7.1
easyocr==1.7.2
fastapi==0.118.2
filelock==3.20.0
fonttools==4.60.1
fpdf2==2.8.4
fsspec==2025.9.0
h11==0.16.0
httptools==0.6.4
huggingface-hub==0.35.3
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
joblib==1.5.1
lazy_loader==0.4
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
ninja==1.13.0
numpy==2.2.6
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
pyclipper==1.3.0.post6
pydantic==2.12.0
pydantic_core==2.41.1
python-bidi==0.6.6
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
shapely==2.1.2
sniffio==1.3.1
starlette==0.48.0
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.10.4
tokenizers==0.22.1
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.57.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
watchfiles==1.1.0
websockets==15.0.1
